moved more types
