+++ packages/@coorpacademy-app-review/src/actions/ui/review-status.ts
@@ -1,19 +1,19 @@
-import {UpdateReviewStatusPayload} from '../../types/review-status';
-
 // -----------------------------------------------------------------------------
 
 export const UPDATE_REVIEW_STATUS = '@@review-status/UPDATE_REVIEW_STATUS';
 
 // -----------------------------------------------------------------------------
 
+export type FinishedStatus = 'finished';
+
 export type UpdateReviewStatus = {
   type: typeof UPDATE_REVIEW_STATUS;
-  payload: UpdateReviewStatusPayload;
+  payload: FinishedStatus;
 };
 
 // -----------------------------------------------------------------------------
 
-export const updateReviewStatus = (payload: UpdateReviewStatusPayload): UpdateReviewStatus => ({
+export const updateReviewStatus = (payload: FinishedStatus): UpdateReviewStatus => ({
   type: UPDATE_REVIEW_STATUS,
   payload
 });
