+++ packages/@coorpacademy-app-review/src/types/common.ts
@@ -16,3 +16,8 @@ export type JWT = {
   usage: string;
   host: string;
 };
+
+export type WithRequired<T, K extends keyof T> = T & {
+  // the "-" is a Mapping Modifier, removes optionality from a prop
+  [P in K]-?: T[P];
+};
