--- packages/@coorpacademy-app-review/src/types/finished-slides.ts
@@ -1,1 +1,0 @@
-export type UpdateFinishedSlidesPayload = {slideNumber: number; value: true};
