--- packages/@coorpacademy-app-review/src/types/review-status.ts
@@ -1,1 +1,0 @@
-export type UpdateReviewStatusPayload = 'finished';
