--- packages/@coorpacademy-app-review/src/types/utils.ts
@@ -1,4 +1,0 @@
-export type WithRequired<T, K extends keyof T> = T & {
-  // the "-" is a Mapping Modifier, removes optionality from a prop
-  [P in K]-?: T[P];
-};
