+++ packages/@coorpacademy-app-review/src/types/views/slides.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import {Dispatchers} from './dispatchers';
-import {StoreState} from './store-state';
+import {Dispatchers} from '../../actions';
+import {StoreState} from '../../configure-store';
 
 type SlidesViewStaticProps = {
   slides: StoreState['data']['slides'];
